fix: build fail
rebase 이후 빌드 안되는 문제 해결
- 중간에 tracking 안되는 파일들 내용 반영
- 필터 태그 정상화
