+++ External/Include/Common/extern.cpp
@@ -6,14 +6,16 @@ GlobalData g_Data = {};
 
 const char* ASSET_TYPE_STRING[] =
 {
-    "MESH",
-    "MESH_DATA",
-    "TEXTURE",
-    "SOUND",
-    "PREFAB",
-    "FLIPBOOK",
-    "SPRITE",
-    "MATERIAL",
-    "GRAPHIC_SHADER",
-    "COMPUTE_SHADER",
+	"MESH",
+	"MESH_DATA",
+	"TEXTURE",
+	"SOUND",
+	"PREFAB",
+	"FLIPBOOK",
+	"ANIMATION",
+	"SKELETON",
+	"SPRITE",
+	"MATERIAL",
+	"GRAPHIC_SHADER",
+	"COMPUTE_SHADER",
 };
