--- Source/Client/AnimationUI.cpp
@@ -1,2 +1,0 @@
-﻿#include "pch.h"
-#include "AnimationUI.h"
