+++ Source/Client/UI/Private/Component/Light3DUI.cpp
@@ -1,7 +1,7 @@
 ﻿#include "pch.h"
-#include "Light3DUI.h"
-
-#include <Engine/CLight3D.h>
+#include "UI/Public/Component/Light3DUI.h"
+#include "Runtime/Public/Actor/CGameObject.h"
+#include "Runtime/Public/Component/Light/CLight3D.h"
 
 Light3DUI::Light3DUI()
 	: ComponentUI("Light3DUI", COMPONENT_TYPE::LIGHT3D)
@@ -24,25 +24,25 @@ void Light3DUI::Render_Update()
 	Vec3 Color = Light3DInfo.info.vColor;
 	ImGui::Text("Color widget:");
 	ImGui::SameLine();
-	ImGui::ColorPicker4("MyColor##1", (float*)&Color);  // RGB 값 편집
+	ImGui::ColorPicker4("MyColor##1", (float*)&Color); // RGB 값 편집
 	pLight3D->SetLightColor(Color);
 
 	// SpecCoeff
 	float SpecCoeff = Light3DInfo.info.SpecCoeff;
 	ImGui::Text("SpecCoeff");
 	ImGui::SameLine();
-	ImGui::InputFloat("##SpecCoeff", &SpecCoeff);  // RGB 값 편집
+	ImGui::InputFloat("##SpecCoeff", &SpecCoeff); // RGB 값 편집
 	pLight3D->SetSpecularCoefficient(SpecCoeff);
 
 	// Ambient
 	Vec3 Ambient = Light3DInfo.info.vAmbient;
 	ImGui::Text("Ambient");
 	ImGui::SameLine();
-	ImGui::InputFloat3("##Ambient", Ambient);  // RGB 값 편집
+	ImGui::InputFloat3("##Ambient", Ambient); // RGB 값 편집
 	pLight3D->SetAmbient(Ambient);
 
 	// LightType
-	const char* szLightType[3] = { "Directional", "Point", "Spot" };
+	const char* szLightType[3] = {"Directional", "Point", "Spot"};
 	int LightType = (int)Light3DInfo.Type;
 	ImGui::Combo("##combo", &LightType, szLightType, 3);
 	if (LightType != (int)Light3DInfo.Type)
