+++ Source/Client/UI/Public/Asset/AnimationUI.h
@@ -1,7 +1,7 @@
 ﻿#pragma once
 #include "AssetUI.h"
-class SkeletonUI :
-    public AssetUI
+
+class AnimationUI :
+	public AssetUI
 {
 };
-
