+++ Source/Client/UI/Public/Asset/SkeletonUI.h
@@ -1,7 +1,7 @@
 ﻿#pragma once
 #include "AssetUI.h"
-class AnimationUI :
-    public AssetUI
+
+class SkeletonUI :
+	public AssetUI
 {
 };
-
