+++ Source/Engine/System/Private/Manager/CAssetMgr.cpp
@@ -2,7 +2,7 @@
 #include "System/Public/Manager/CAssetMgr.h"
 
 CAssetMgr::CAssetMgr()
-    : m_bAssetChanged(false)
+	: m_bAssetChanged(false)
 {
 }
 
@@ -13,87 +13,87 @@ CAssetMgr::~CAssetMgr()
 Ptr<CTexture> CAssetMgr::CreateTexture(const wstring& _Key, UINT _Width, UINT _Height,
                                        DXGI_FORMAT _PixelFormat, UINT _BindFlag, D3D11_USAGE _Usage)
 {
-    Ptr<CTexture> pTex = FindAsset<CTexture>(_Key);
+	Ptr<CTexture> pTex = FindAsset<CTexture>(_Key);
 
-    if (nullptr != pTex)
-        return pTex;
+	if (nullptr != pTex)
+		return pTex;
 
-    pTex = new CTexture;
-    if (FAILED(pTex->Create(_Width, _Height, _PixelFormat, _BindFlag, _Usage)))
-    {
-        assert(nullptr);
-        return nullptr;
-    }
+	pTex = new CTexture;
+	if (FAILED(pTex->Create(_Width, _Height, _PixelFormat, _BindFlag, _Usage)))
+	{
+		assert(nullptr);
+		return nullptr;
+	}
 
-    pTex->SetKey(_Key);
-    m_mapAsset[static_cast<UINT>(ASSET_TYPE::TEXTURE)].insert(make_pair(_Key, pTex.Get()));
-    m_bAssetChanged = true;
-    return pTex;
+	pTex->SetKey(_Key);
+	m_mapAsset[static_cast<UINT>(ASSET_TYPE::TEXTURE)].insert(make_pair(_Key, pTex.Get()));
+	m_bAssetChanged = true;
+	return pTex;
 }
 
 Ptr<CTexture> CAssetMgr::CreateTexture(const wstring& _Key, ComPtr<ID3D11Texture2D> _Tex2D)
 {
-    Ptr<CTexture> pTex = FindAsset<CTexture>(_Key);
+	Ptr<CTexture> pTex = FindAsset<CTexture>(_Key);
 
-    if (nullptr != pTex)
-        return pTex;
+	if (nullptr != pTex)
+		return pTex;
 
-    pTex = new CTexture;
-    if (FAILED(pTex->Create(_Tex2D)))
-    {
-        assert(nullptr);
-        return nullptr;
-    }
+	pTex = new CTexture;
+	if (FAILED(pTex->Create(_Tex2D)))
+	{
+		assert(nullptr);
+		return nullptr;
+	}
 
-    pTex->SetKey(_Key);
-    m_mapAsset[static_cast<UINT>(ASSET_TYPE::TEXTURE)].insert(make_pair(_Key, pTex.Get()));
-    m_bAssetChanged = true;
-    return pTex;
+	pTex->SetKey(_Key);
+	m_mapAsset[static_cast<UINT>(ASSET_TYPE::TEXTURE)].insert(make_pair(_Key, pTex.Get()));
+	m_bAssetChanged = true;
+	return pTex;
 }
 
 void CAssetMgr::GetAssetNames(ASSET_TYPE _Type, vector<wstring>& _vecAssetNames)
 {
-    for (const auto& pair : m_mapAsset[static_cast<UINT>(_Type)])
-    {
-        _vecAssetNames.push_back(pair.first);
-    }
+	for (const auto& pair : m_mapAsset[static_cast<UINT>(_Type)])
+	{
+		_vecAssetNames.push_back(pair.first);
+	}
 }
 
 void CAssetMgr::DeleteAsset(ASSET_TYPE _Type, const wstring& _Key)
 {
-    map<wstring, Ptr<CAsset>>& mapAsset = m_mapAsset[static_cast<UINT>(_Type)];
+	map<wstring, Ptr<CAsset>>& mapAsset = m_mapAsset[static_cast<UINT>(_Type)];
 
-    auto iter = mapAsset.find(_Key);
+	auto iter = mapAsset.find(_Key);
 
-    if (mapAsset.end() == iter)
-        return;
+	if (mapAsset.end() == iter)
+		return;
 
-    mapAsset.erase(iter);
+	mapAsset.erase(iter);
 
-    m_bAssetChanged = true;
+	m_bAssetChanged = true;
 }
 
 Ptr<CMeshData> CAssetMgr::LoadFBX(const wstring& _strPath)
 {
-    wstring strFileName = path(_strPath).stem();
+	wstring strFileName = path(_strPath).stem();
 
-    wstring strName = L"MeshData\\";
-    strName += strFileName + L".mdat";
+	wstring strName = L"MeshData\\";
+	strName += strFileName + L".mdat";
 
-    Ptr<CMeshData> pMeshData = FindAsset<CMeshData>(strName);
+	Ptr<CMeshData> pMeshData = FindAsset<CMeshData>(strName);
 
-    if (nullptr != pMeshData)
-        return pMeshData;
+	if (nullptr != pMeshData)
+		return pMeshData;
 
-    pMeshData = CMeshData::LoadFromFBX(_strPath);
-    pMeshData->SetKey(strName);
-    pMeshData->SetRelativePath(strName);
+	pMeshData = CMeshData::LoadFromFBX(_strPath);
+	pMeshData->SetKey(strName);
+	pMeshData->SetRelativePath(strName);
 
-    m_mapAsset[static_cast<UINT>(ASSET_TYPE::MESH_DATA)].
-        insert(make_pair(strName, pMeshData.Get()));
+	m_mapAsset[static_cast<UINT>(ASSET_TYPE::MESH_DATA)].
+		insert(make_pair(strName, pMeshData.Get()));
 
-    // meshdata 를 실제파일로 저장
-    pMeshData->Save(strName);
+	// meshdata 를 실제파일로 저장
+	pMeshData->Save(strName);
 
-    return pMeshData;
+	return pMeshData;
 }
