+++ Source/Engine/System/Private/Rendering/Tool/FBX/CFBXLoader.cpp
@@ -3,8 +3,6 @@
 #include "System/Public/Manager/CAssetMgr.h"
 #include "System/Public/Manager/CPathMgr.h"
 
-class CTexture;
-
 CFBXLoader::CFBXLoader()
 	: m_pManager(nullptr)
 	  , m_pScene(nullptr)
@@ -36,7 +34,7 @@ CFBXLoader::~CFBXLoader()
 	}
 }
 
-void CFBXLoader::init()
+void CFBXLoader::Init()
 {
 	m_pManager = FbxManager::Create();
 	if (nullptr == m_pManager)
@@ -361,7 +359,7 @@ Vec4 CFBXLoader::GetMtrlData(FbxSurfaceMaterial* _pSurface
 		dFactor = tMtrlFactorProperty.Get<FbxDouble>();
 	}
 
-	auto vRetVal = Vec4(static_cast<float>(vMtrl.mData[0]) * static_cast<float>(dFactor),
+	Vec4 vRetVal = Vec4(static_cast<float>(vMtrl.mData[0]) * static_cast<float>(dFactor),
 	                    static_cast<float>(vMtrl.mData[1]) * static_cast<float>(dFactor),
 	                    static_cast<float>(vMtrl.mData[2]) * static_cast<float>(dFactor),
 	                    static_cast<float>(dFactor));
@@ -516,8 +514,6 @@ void CFBXLoader::CreateMaterial()
 
 void CFBXLoader::LoadSkeleton(FbxNode* _pNode)
 {
-	int iChildCount = _pNode->GetChildCount();
-
 	LoadSkeleton_Re(_pNode, 0, 0, -1);
 }
 
@@ -527,7 +523,7 @@ void CFBXLoader::LoadSkeleton_Re(FbxNode* _pNode, int _iDepth, int _iIdx, int _i
 
 	if (pAttr && pAttr->GetAttributeType() == FbxNodeAttribute::eSkeleton)
 	{
-		auto pBone = new tBone;
+		tBone* pBone = new tBone;
 
 		string strBoneName = _pNode->GetName();
 
@@ -553,15 +549,13 @@ void CFBXLoader::LoadAnimationClip()
 	{
 		FbxAnimStack* pAnimStack = m_pScene->FindMember<FbxAnimStack>(m_arrAnimName[i]->Buffer());
 
-
 		//FbxAnimEvaluator* pevaluator = m_pScene->GetAnimationEvaluator();
 		//m_pScene->SetCurrentAnimationStack();
-
 
 		if (!pAnimStack)
 			continue;
 
-		auto pAnimClip = new tAnimClip;
+		tAnimClip* pAnimClip = new tAnimClip;
 
 		string strClipName = pAnimStack->GetName();
 		pAnimClip->strName = wstring(strClipName.begin(), strClipName.end());
@@ -613,7 +607,7 @@ void CFBXLoader::LoadAnimationData(FbxMesh* _pMesh, tContainer* _pContainer)
 	// Skin 개수만큼 반복을하며 읽는다.
 	for (int i = 0; i < iSkinCount; ++i)
 	{
-		auto pSkin = static_cast<FbxSkin*>(_pMesh->GetDeformer(i, FbxDeformer::eSkin));
+		FbxSkin* pSkin = static_cast<FbxSkin*>(_pMesh->GetDeformer(i, FbxDeformer::eSkin));
 
 		if (pSkin)
 		{
@@ -623,6 +617,8 @@ void CFBXLoader::LoadAnimationData(FbxMesh* _pMesh, tContainer* _pContainer)
 				// Cluster 를 얻어온다
 				// Cluster == Joint == 관절
 				int iClusterCount = pSkin->GetClusterCount();
+
+				FbxAMatrix matNodeTransform = GetTransform(_pMesh->GetNode());
 
 				for (int j = 0; j < iClusterCount; ++j)
 				{
@@ -636,10 +632,12 @@ void CFBXLoader::LoadAnimationData(FbxMesh* _pMesh, tContainer* _pContainer)
 					if (-1 == iBoneIdx)
 						assert(NULL);
 
-					FbxAMatrix matNodeTransform = GetTransform(_pMesh->GetNode());
-
 					// Weights And Indices 정보를 읽는다.
 					LoadWeightsAndIndices(pCluster, iBoneIdx, _pContainer);
+
+					// 한 번 읽은 bone 정보는 더 이상 읽지 않는다.
+					if (m_vecBone[iBoneIdx]->vecKeyFrame.size() > 0)
+						continue;
 
 					// Bone 의 OffSet 행렬 구한다.
 					LoadOffsetMatrix(pCluster, matNodeTransform, iBoneIdx, _pContainer);
@@ -724,31 +722,41 @@ void CFBXLoader::LoadKeyframeTransform(FbxNode* _pNode, FbxCluster* _pCluster
 	matReflect.mData[2] = v3;
 	matReflect.mData[3] = v4;
 
-	m_vecBone[_iBoneIdx]->matBone = _matNodeTransform;
+	m_vecBone[_iBoneIdx]->vecKeyFrame.resize(m_vecAnimClip.size());
 
 	//FbxTime::EMode eTimeMode = m_pScene->GetGlobalSettings().GetTimeMode();
 	FbxTime::EMode eTimeMode = FbxTime::EMode::eFrames30;
 
-	FbxLongLong llStartFrame = m_vecAnimClip[0]->tStartTime.GetFrameCount(eTimeMode);
-	FbxLongLong llEndFrame = m_vecAnimClip[0]->tEndTime.GetFrameCount(eTimeMode);
-
-	for (FbxLongLong i = llStartFrame; i < llEndFrame; ++i)
+	// 애니메이션 clip을 각각 추출
+	for (int clipIdx = 0; clipIdx < m_vecAnimClip.size(); ++clipIdx)
 	{
-		tKeyFrame tFrame = {};
-		FbxTime tTime = 0;
+		FbxAnimStack* pCurAnimStack = m_pScene->FindMember<FbxAnimStack>(m_arrAnimName[clipIdx]->Buffer());
+		m_pScene->SetCurrentAnimationStack(pCurAnimStack);
 
-		tTime.SetFrame(i, eTimeMode);
+		FbxLongLong llStartFrame = m_vecAnimClip[clipIdx]->tStartTime.GetFrameCount(eTimeMode);
+		FbxLongLong llEndFrame = m_vecAnimClip[clipIdx]->tEndTime.GetFrameCount(eTimeMode);
 
-		FbxAMatrix matFromNode = _pNode->EvaluateGlobalTransform(tTime) * _matNodeTransform;
-		FbxAMatrix matCurTrans = matFromNode.Inverse() * _pCluster->GetLink()->
-		                                                            EvaluateGlobalTransform(tTime);
-		matCurTrans = matReflect * matCurTrans * matReflect;
+		for (FbxLongLong i = llStartFrame; i < llEndFrame; ++i)
+		{
+			tKeyFrame tFrame = {};
+			FbxTime tTime = 0;
 
-		tFrame.dTime = tTime.GetSecondDouble();
-		tFrame.matTransform = matCurTrans;
+			tTime.SetFrame(i, eTimeMode);
 
-		m_vecBone[_iBoneIdx]->vecKeyFrame.push_back(tFrame);
+			FbxAMatrix matFromNode = _pNode->EvaluateGlobalTransform(tTime) * _matNodeTransform;
+			FbxAMatrix matCurTrans = matFromNode.Inverse() * _pCluster->GetLink()->
+			                                                            EvaluateGlobalTransform(tTime);
+			matCurTrans = matReflect * matCurTrans * matReflect;
+
+			tFrame.dTime = tTime.GetSecondDouble();
+			tFrame.matTransform = matCurTrans;
+
+			m_vecBone[_iBoneIdx]->vecKeyFrame[clipIdx].push_back(tFrame);
+		}
 	}
+
+	FbxAnimStack* pCurAnimStack = m_pScene->FindMember<FbxAnimStack>(m_arrAnimName[0]->Buffer());
+	m_pScene->SetCurrentAnimationStack(pCurAnimStack);
 }
 
 void CFBXLoader::LoadOffsetMatrix(FbxCluster* _pCluster
@@ -778,6 +786,7 @@ void CFBXLoader::LoadOffsetMatrix(FbxCluster* _pCluster
 	matOffset = matReflect * matOffset * matReflect;
 
 	m_vecBone[_iBoneIdx]->matOffset = matOffset;
+	m_vecBone[_iBoneIdx]->matBone = _matNodeTransform;
 }
 
 
@@ -804,7 +813,7 @@ void CFBXLoader::LoadWeightsAndIndices(FbxCluster* _pCluster
 
 int CFBXLoader::FindBoneIndex(string _strBoneName)
 {
-	auto strBoneName = wstring(_strBoneName.begin(), _strBoneName.end());
+	wstring strBoneName = wstring(_strBoneName.begin(), _strBoneName.end());
 
 	for (UINT i = 0; i < m_vecBone.size(); ++i)
 	{
