+++ Source/Engine/System/Public/Rendering/Shader/CWeightMapCS.h
@@ -1,6 +1,7 @@
 ﻿#pragma once
 #include "System/Public/Rendering/Shader/CComputeShader.h"
 
+class CTexture;
 class CStructuredBuffer;
 
 class CWeightMapCS :
