--- Source/Engine/extern.cpp
@@ -1,22 +1,0 @@
-﻿#include "pch.h"
-#include "Common/global.h"
-
-TransInfo g_Trans = {};
-GlobalData g_Data = {};
-
-
-const char* ASSET_TYPE_STRING[] =
-{
-    "MESH",
-    "MESH_DATA",
-    "TEXTURE",
-    "SOUND",
-    "PREFAB",
-    "FLIPBOOK",
-	"ANIMATION",
-	"SKELETON",
-    "SPRITE",
-    "MATERIAL",
-    "GRAPHIC_SHADER",
-    "COMPUTE_SHADER",
-};
